backend/backend_requirements.txt
flask
psycopg2-binary
python-dotenv
flask-sqlalchemy
PyYAML